rename package api-utils to eoepca-api-utils, and added the needed dependencies for it
