+++ check_manager/pyproject.toml
@@ -15,7 +15,7 @@ dependencies = [
     "httpx>=0.27.2",
     "jsonschema>=4.23.0",
     "kubernetes_asyncio>=31.1.0",
-    "api-utils>=0.1.0",
+    "eoepca-api-utils>=0.1.0",
     "plugin-utils>=0.1.1",
     "pydantic>=2.9.2",
     "pydantic-settings>=2.8.1",
@@ -34,7 +34,7 @@ dev-dependencies = [
 ]
 
 [tool.uv.sources]
-api-utils = { git = "https://github.com/EOEPCA/resource-health.git", subdirectory = "api-utils" }
+eoepca-api-utils = { git = "https://github.com/EOEPCA/resource-health.git", subdirectory = "eoepca-api-utils" }
 plugin_utils = { git = "https://github.com/EOEPCA/resource-health.git", subdirectory = "plugin-utils" }
 eoepca-security = { git = "https://github.com/EOEPCA/python-eoepca-security.git", branch = "2.0.0-rc2" }
 
@@ -47,7 +47,6 @@ asyncio_default_fixture_loop_scope = "function"
 
 [tool.mypy]
 python_version = "3.12"
-packages = "check_backends"
 # Enable later
 # disallow_untyped_defs = true
 show_error_context = true
