+++ eoepca-api-utils/README.md
@@ -1,4 +1,4 @@
-# Api Utils
+# EOEPCA+ Api Utils
 
 Name is subject to change.
