+++ eoepca-api-utils/pyproject.toml
@@ -1,13 +1,18 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
 [project]
-name = "api-utils"
+name = "eoepca-api-utils"
 version = "0.1.0"
 description = "FastAPI utilities, such as exceptions and JSON:API types"
 readme = "README.md"
 requires-python = ">=3.12"
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
+dependencies = [
+    "fastapi[standard]>=0.115.4",
+    "pydantic>=2.9.2",
+    "pydantic-settings>=2.8.1",
+]
 
 [tool.uv]
 dev-dependencies = [
@@ -17,12 +22,11 @@ dev-dependencies = [
 ]
 
 # [tool.pytest.ini_options]
-# addopts = "-ra -q --import-mode=importlib --cov=api_utils --cov-report=term-missing"
+# addopts = "-ra -q --import-mode=importlib --cov=eoepca_api_utils --cov-report=term-missing"
 # testpaths = ["tests"]
 
 [tool.mypy]
 python_version = "3.12"
-packages = "api_utils"
 disallow_untyped_defs = true
 show_error_context = true
 pretty = true
