+++ plugin-utils/dummy.py
@@ -1,0 +1,9 @@
+def foo(exceptions: type[BaseException] | tuple[type[BaseException], ...]) -> None:
+    try:
+        raise TypeError("Bad type :(")
+    except exceptions:
+        print("Exception occured")
+    
+
+foo(TypeError)
+foo((ValueError,))
